chore: add new mapped url for testnet and renamed mapping functions
