+++ apps/web/src/components/EntityCards/AssetCard/index.tsx
@@ -9,7 +9,7 @@ import { HorizontalLine } from 'components/HorizontalLine'
 import { Tag } from 'components'
 import { getEntityIcon } from 'utils/getEntityIcon'
 import { useNavigate } from 'react-router-dom'
-import { getCDNURL } from '@ixo-webclient/utils'
+import { getMappedCDNURL } from '@ixo-webclient/utils'
 
 interface Props extends HTMLFlexBoxProps {
   id?: string
@@ -60,7 +60,7 @@ export const AssetCard: React.FC<Props> = ({
     >
       <FlexBox
         position='relative'
-        background={`url(${getCDNURL(cardImage)})`}
+        background={`url(${getMappedCDNURL(cardImage)})`}
         width='100%'
         height='200px'
         $backgroundSize='cover'
